requirements.txt
# DCASE 2025 Task 3 Sound Event Localization and Detection Requirements
# 
# This file contains the essential dependencies for the DCASE 2025 Task 3 SELD project.
# Install all dependencies with: pip install -r requirements.txt

# Core Deep Learning Frameworks
torch>=1.13.0
torchaudio>=0.13.0
torchvision>=0.14.0

# Audio Processing
librosa>=0.9.2
soundfile>=0.11.0

# Data Processing and Numerical Computing
numpy>=1.21.0
scipy>=1.8.0
pandas>=1.4.0

# Computer Vision (for model components)
opencv-python>=4.6.0
Pillow>=9.0.0

# Progress and Visualization
tqdm>=4.64.0
matplotlib>=3.5.0

# Logging and Monitoring
tensorboard>=2.9.0

# Model Analysis Tools
ptflops>=0.6.6
thop>=0.1.0

# Scientific Computing
scikit-learn>=1.1.0

# System and OS utilities
psutil>=5.9.0

# Testing and Development
pytest>=7.0.0
pytest-cov>=3.0.0

# Development Tools
ipython>=8.0.0
jupyter>=1.0.0

# Optional: For advanced model implementations
transformers>=4.20.0
einops>=0.4.0

# Optional: For visualization
seaborn>=0.11.0
plotly>=5.0.0

# Optional: For experiment tracking
wandb>=0.13.0
mlflow>=2.0.0

# Optional: For audio format support
ffmpeg-python>=0.2.0
pydub>=0.25.0

# Optional: For file handling
h5py>=3.6.0

# Optional: For configuration management
hydra-core>=1.2.0
omegaconf>=2.2.0
pyyaml>=6.0.0

# Optional: For hyperparameter optimization
optuna>=3.0.0

# Optional: For advanced numpy operations
numba>=0.56.0

# Optional: For memory profiling
memory-profiler>=0.60.0

# Optional: For GPU monitoring
gpustat>=1.0.0
nvidia-ml-py3>=7.352.0

# Optional: For audio streaming
pyaudio>=0.2.11

# Optional: For model profiling
torch-tb-profiler>=0.4.0

# Optional: For advanced audio analysis
essentia>=2.1b6

# Optional: For code formatting
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0

# Optional: For environment management
python-dotenv>=0.20.0

# Optional: For logging
loguru>=0.6.0

# Optional: For testing
coverage>=6.0.0

# Optional: For code quality
mypy>=0.971
pylint>=2.14.0

# Optional: For model optimization
pytorch-lightning>=1.7.0 